+++ api-user/src/main/kotlin/zipbap/user/api/file/service/FileService.kt
@@ -1,4 +1,4 @@
-package zipbap.user.api.recipe.service
+package zipbap.user.api.file.service
 
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
@@ -6,13 +6,15 @@ import zipbap.global.domain.file.FileEntity
 import zipbap.global.domain.file.FileRepository
 import zipbap.global.domain.file.FileStatus
 import zipbap.global.domain.recipe.Recipe
+import zipbap.global.domain.user.User
+import kotlin.collections.forEach
 
 /**
  * AOP를 위해 따로 Service 작성
  */
 @Service
 @Transactional
-class RecipeFileService(
+class FileService(
         private val fileRepository: FileRepository
 ) {
     /**
@@ -20,11 +22,11 @@ class RecipeFileService(
      * - 사용되지 않는 파일 → UNTRACKED
      * - 사용된 파일 → 주어진 targetStatus 로 업데이트
      */
-    fun updateFileStatuses(
-            recipeId: String,
-            usedFileUrls: Set<String>,
-            targetStatus: FileStatus,
-            recipe: Recipe
+    fun updateRecipeFileStatuses(
+        recipeId: String,
+        usedFileUrls: Set<String>,
+        targetStatus: FileStatus,
+        recipe: Recipe
     ) {
         // 기존에 연결된 파일들
         val attachedFiles = fileRepository.findAllByRecipeId(recipeId)
@@ -58,6 +60,33 @@ class RecipeFileService(
             file.status = FileStatus.UNTRACKED
             file.recipe = null
             fileRepository.save(file)
+        }
+    }
+
+    /**
+     * 공통 파일 상태 업데이트
+     * - 사용 안 된 파일 → UNTRACKED + user = null
+     * - 사용된 파일 → targetStatus + user = managedUser
+     */
+    fun updateUserFileStatuses(
+        usedFileUrls: Set<String>,
+        targetStatus: FileStatus,
+        managedUser: User
+    ) {
+        val attachedFiles = fileRepository.findAllByUser(managedUser)
+
+        // 사용 안된 파일 정리
+        attachedFiles.filter { it.fileUrl !in usedFileUrls }.forEach { file ->
+            file.status = FileStatus.UNTRACKED
+            file.user = null
+        }
+
+        // 사용된 파일 상태 처리
+        usedFileUrls.forEach { url ->
+            fileRepository.findByFileUrl(url)?.apply {
+                this.user = managedUser   // ✅ Detached user 대신 managedUser 사용
+                this.status = targetStatus
+            }
         }
     }
 }
